Adds better operate functionality
Now, if you type an operand, then a operator, then
an operand, and finally another operator, it will
evaluate the initial expression, save it as the
first operand, and let you continue with the
expression. Also, if you evaluate an expression,
then type a number, it will replace the result of
the previous expression, but, if you instead type
an operator, it will treat the result of the
previous expression as the first operand of a new
expression.